Replace usage of new UnableToDecryptEventEntity by usage of existing EventInsertEntity
