--- matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/UnableToDecryptEventLiveObserver.kt
@@ -1,83 +1,0 @@
-/*
- * Copyright 2022 The Matrix.org Foundation C.I.C.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.matrix.android.sdk.internal.database
-
-import com.zhuinden.monarchy.Monarchy
-import io.realm.RealmConfiguration
-import io.realm.RealmResults
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.sync.Mutex
-import kotlinx.coroutines.sync.withLock
-import org.matrix.android.sdk.internal.database.mapper.asDomain
-import org.matrix.android.sdk.internal.database.model.EventEntity
-import org.matrix.android.sdk.internal.database.model.UnableToDecryptEventEntity
-import org.matrix.android.sdk.internal.database.model.UnableToDecryptEventEntityFields
-import org.matrix.android.sdk.internal.database.query.where
-import org.matrix.android.sdk.internal.di.SessionDatabase
-import org.matrix.android.sdk.internal.session.UnableToDecryptEventLiveProcessor
-import timber.log.Timber
-import javax.inject.Inject
-
-internal class UnableToDecryptEventLiveObserver @Inject constructor(
-        @SessionDatabase realmConfiguration: RealmConfiguration,
-        private val processors: Set<@JvmSuppressWildcards UnableToDecryptEventLiveProcessor>
-) :
-        RealmLiveEntityObserver<UnableToDecryptEventEntity>(realmConfiguration) {
-
-    private val lock = Mutex()
-
-    override val query = Monarchy.Query {
-        it.where(UnableToDecryptEventEntity::class.java)
-    }
-
-    override fun onChange(results: RealmResults<UnableToDecryptEventEntity>) {
-        observerScope.launch {
-            lock.withLock {
-                if (!results.isLoaded || results.isEmpty()) {
-                    return@withLock
-                }
-                val copiedEvents = ArrayList<UnableToDecryptEventEntity>(results.size)
-                Timber.v("UnableToDecryptEventEntity updated with ${results.size} results in db")
-                results.forEach {
-                    // don't use copy from realm over there
-                    val copiedEvent = UnableToDecryptEventEntity(eventId = it.eventId)
-                    copiedEvents.add(copiedEvent)
-                }
-                awaitTransaction(realmConfiguration) { realm ->
-                    Timber.v("##Transaction: There are ${copiedEvents.size} events to process ")
-                    copiedEvents.forEach { utdEvent ->
-                        val eventId = utdEvent.eventId
-                        val event = EventEntity.where(realm, eventId).findFirst()
-                        if (event == null) {
-                            Timber.v("Event $eventId not found")
-                            return@forEach
-                        }
-                        val domainEvent = event.asDomain()
-                        processors.forEach {
-                            it.process(realm, domainEvent)
-                        }
-                    }
-                    realm.where(UnableToDecryptEventEntity::class.java)
-                            .`in`(UnableToDecryptEventEntityFields.EVENT_ID, copiedEvents.map { it.eventId }.toTypedArray())
-                            .findAll()
-                            .deleteAllFromRealm()
-                }
-                processors.forEach { it.onPostProcess() }
-            }
-        }
-    }
-}
