--- matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/UnableToDecryptEventLiveProcessor.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2022 The Matrix.org Foundation C.I.C.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.matrix.android.sdk.internal.session
-
-import io.realm.Realm
-import org.matrix.android.sdk.api.session.events.model.Event
-
-internal interface UnableToDecryptEventLiveProcessor {
-
-    /**
-     * Process the given event.
-     * @param realm a realm instance
-     * @param event the event to be processed
-     * @return true if it has been processed, false if it was ignored.
-     */
-    fun process(realm: Realm, event: Event): Boolean
-
-    /**
-     * Called after transaction.
-     * Maybe you prefer to process the events outside of the realm transaction.
-     */
-    suspend fun onPostProcess() {
-        // Noop by default
-    }
-}
